Reorder build to fix dependency issue. Was getting error: info: Build ['-q', 'install']: /DATA/ARCHIVED_1/mah131/svn/ASKAPsoft/Code/Components/Analysis/analysis/trunk/build.py [sourcefitting/RadioSource.cc:36:53: error: evaluationutilities/EvaluationUtilities.h: No such file or directory
Need to to do evaluation first so that we make the include file available
for analysis/trunk/sourcefitting/RadioSource.cc

Changeset:6240 (2008-07-17) introduced this dependency which is issue
for the full test builds.  The rbuild command (provided dependencies
are listed correctly) should take care of this for developer builds.
All the more reason to get automated testing going to catch them.
Without doing the full clean rebuild these sort of issues are hard to catch.