Added two classes and unit tests. These classes are an encapsulated representation of the TOS metadata. This is designed to provide an abstraction above the raw ICE TimeTaggedTypedValueMap type.
The purpose behind this is that a converter to/from TimeTaggedTypedValueMap and TosMetadata will be created. This means code such as the correlator simulator and the ingest pipeline are not exposed to the ICE type (TimeTaggedTypedValueMap) which is very abstract and requires a fair bit of boiler plate code to populate.