ASKAPSDP-1634 - Further improvement of the model components factory method and the related constructors.
I've made a change to the ModelFactor code to better use the 
shared_ptr class, using the reset() function. I've updated the 
constructors for each model component class to allow 
specification of all parameters at once, to simplify the 
interface. I've also tried to add the const keyword in as many 
places as needed, and general formatting improvements.