ASKAPSDP-1629 - Improving the merging step for edge sources.
This change allows us to partially speed up the merging of the edge 
source list, by using the mergeIntoList function from Duchamp to merge 
new edge sources into the list if need be. We still need the complete 
merging step to ensure a unique list, but this process now runs 
quicker - for a large spectral-line test case it halved the time taken 
for merging the list.