ASKAPSDP-1196 - Further fixes to the RM Synthesis code following code review CR-ASKAPSDP-1.
This adds a further fix to the defineVectors() function to avoid the 
problem Max identified of assigning a vector to something already 
defined to a different length. 
It also addresses a couple of issues Ben identified: the defaultWeight 
constant is now defined as static in the .cc file; and the 
CPPUNIT_ASSERT_*_EQUAL calls are changed so that the expected value is 
correctly located as the first argument.