ASKAPSDP-1646: Fixes a bug in casdaupload where absolute paths did not work
The config file allows users to specify the input artifacts as either absolute
or relative paths, or just a filename only. For absolute paths, when tarring
files, it is necessary to pass the "--directory" option to the tar command
otherwise the full path will be included in the tarfile. This commit implements
that fix.