ASKAPSDP-1693 - Initial commit of HI emission-line catalogue
This changeset introduces classes to cover CASDA-compatible HI 
emission-line objects and associated catalogues, as well as a 
test/demonstration application that can produce a catalogue from the 
Duchamp verification cube. The necessary fields are all present, 
although the calculations are not fully implemented, so the bulk of 
the information in the catalogues are zero.