ASKAPSDP-1477 - enabling restored image output in simager
This changeset enables simager to write out the restored image, along 
with the psf.image (preconditioned psf). Some slight modifications 
were made to the way the params are handled, as the restore solver 
overwrites the image.slice parameter, meaning you couldn't send both 
the clean model and the restored image to the master for writing. 
The image.slice param is now copied to model.slice before running the 
restore solver, so that all images can be written.