ASKAPSDP-1477 - Adding flux units and rest frequency to the cubes
This change enables the writing of the flux units to the cubes - these 
default to Jy/pixel, but are changed to Jy/beam when the restoring 
beam is set. We also allow the setting of the rest frequency in the 
model, restored and residual images, based on an input parameter, in 
the manner of what was used for makecube.