ASKAPSDP-1699 - better spectral-unit handling for CASDA catalogues
This change improves the way we process the spectral units for the 
CASDA catalogues. It introduces a function to change the spectral type 
to the appropriate one for the CASDA catalogue (in this case, FREQ). 
It also updates the units to those required by CASDA. This has been 
utilised by both the island and component catalogues, so that even 
when an image with VEL type is given, we end up with frequencies.