ASKAPSDP-1576 - Early-science simulation
This change updates the scripts to run the early science simulations. 
I've added a few new parameters, so that we can produce more than one 
final MS (default is 8), and we have better control over the memory 
request for the csimulator jobs. We also set the striping for the 
filesystem before starting anything. The default observing time is 
reduced to 8 hours.