ASKAPSDP-1696 - pipeline scripts
Changing the location of the scripts from SCRIPTDIR to PIPELINEDIR (as 
we will have this in the module when released), and adding some 
testing of the existence of 'lfs' (so that if run on a system other 
than galaxy it will still produce files but not submit them). Also 
adding the header template to all scripts.