ASKAPSDP-1696 - fixes to pipeline scripts
This change fixs a bug with the stats reporting for parallel jobs - we 
now have to explicitly set the number of cpus prior to calling 
extractStats, but it now gets the worker stats correct.

Other changes include adding the version number to the init script, 
and then allowing this to be written out to the screen and the jobs 
list at run time.