First updates to move from casacore 1.6.0 to 2.0.3
Regression tests not done - I just needed to check this in as I was diverging too much.


Interface changes,

A number of small interface changes have been made resulting in changes to almost all of CPPUNIT tests that assert equality. I have put in functional type casts in our tests instead of violating any of the macros

mslist list() method had changed

casacore has of course removed the components so I have too.

casacore overloaded the insertion operator (<<) to deal with container classes. This was already done in AskapUtils so created an ambiguity. This was made “unfixable” without including the casa namespace in everything that used Askaploggers - which I did not want to do. I have patched the casacore version to use the standard namespace for all those operators that operate on standard containers - which at least is compliant with the assumptions of ADL (or Koenig lookup)

This also required the removal of the overload in AskapUtils… and the inclusion of the patches casacore/BasicSL/SLIO.h header to provide the functionality. 

Lots of changes to the locations of header files etc.


A large number of patches are no longer needed are:

casa patches on 1.6.0a - are they needed for 2.0.3

Array::ArrayMath.h & .cc  - No - all the functionality is now in ArrayMath - they are not needed
Array::MaskArrMath.h - Not required - all functionality in 2.03
Array::MaskArrMath.tcc - change to patch required. The patched square method is still required to I have replicated it. Its functionality seems to be to iterate through the array pulling out masks … and the patch is to cover the case where there are no valid elements in the subarray.
Array::ArrayPartMath.h & .tcc — ok so I <think> these are not required as the functionality that the patches provide is now being provided - but in a slightly different manner than the patch suggests. It looks like the interfaces are the same - so I will not use the patches - but beware. The functionals affected are the partialInterHExile and Quartile functions. The new code uses a function …Interfractile that is called with different arguments for the various use cases … perhaps a good thing
BasicSL::Complex.h - I’ve kept it - it explicitly wants to use the norm from the standard namespace and I see no harm in that. - changed my mind - there seems to be no logic as to which functions are moved … i will not in
BasicSL::Complexfwd.h — no longer required 
System::ObjectID.cc — no longer required
ComponentModels::GaussianShape.cc — moved away 
FITS::FITSDateUtil — no longer required
Images::ImageRegrid.tcc — still needed - this is mitch’s bulk coordinate conversion patch. 
LatticeStatistics — as in the Complex.h patch this forces a use of the real and imag from the standard namespace. I am not sure this is needed.
MSConcat — ok two problems here. 1) this file has moved 2) the interface to the concatenate method has entirely changed …. warning
FitGaussian.h - we still need this and it looks like the underlying file has not changed in structure or location so I think we can just apply as is.
FitGaussian.tcc - we still need this but it cannot be applied as is - I will have to replicate it
SerialHelper.h  & cc— no longer needed
FFTServer.hcc — more declaring of std:: functions - this time exp. I suspect this is a compiler problem for someone - I will not use these patches in until i need to.

So the patches that are definitely needed are:

Array::MaskArrMath.tcc
Images::ImageRegrid.tcc
FitGaussian.h
FitGaussian.tcc 

I will add these to the 2.0.3 in 3rd party.