ASKAPSDP-1844 - Updates to fitting unit tests
This change adds a new unit test to the source-fitting, which fits to 
two Gaussians, each with FWHM of 3 pix, separated by 3 pix, with one 
half the flux of the other. The current code fails to fit two 
Gaussians to it, as the initial estimate only finds a single source.