ASKAPSDP-1722 - Ability to fit additional confused components.
This change gives Selavy the ability to fit additional components that 
weren't picked up in the initial estimation. If a fit converges but is 
poor, we subtract off the fitted Gaussians, find the brightest 
residual component, add it to the initial estimates and re-fit.
This change involved a number of changes to the interfaces of the 
Fitter and related tasks. I have also moved Component.{h,cc} to 
SubComponent.{h,cc} (reflecting the name of the class defined 
therein). Various bits of cleaning up of the code were done at the 
same time.
It also necessitated changes to the ComponentImager, so that we could 
directly call the evaluateGaussian function.
The documentation has also been updated.