ASKAPSDP-1955 - Adding CHANGES file to top-level directory.
A file describing changes (bug fixes, new features etc) for each 
release is added to the top-level directory. The plan is that this 
will be updated immediately prior to any release branching or tagging. 
The contents come from release notification emails sent out by Ben 
Humphreys or Matthew Whiting, together with (for earlier releases) a 
list of JIRA (or Redmine) tickets.