ASKAPSDP-2012 - Ensuring thumbnail images are copied.
With the previous changes, the thumbnail images were not being copied. 
I've rejigged the code so that the copy/checksum now operates out of 
the element classes, with a specialisation for the ImageElement to 
take care of the thumbnail images as well as the image file itself.
Tested and appears to work fine.