ASKAPSDP-2021 - Better handling of cellsize parameters
This change makes a better check of the validity of the CELLSIZE & 
CELLSIZE_SPECTRAL parameters. 
It also removes the reliance on the advise capability for 
spectral-line imaging, as simager does not use that. If the 
CELLSIZE_SPECTRAL or NUM_PIXELS_SPECTRAL parameters are invalid, the 
spectral-line imaging job is not run.