ASKAPSDP-2057 & ASKAPSDP-2058 - pipeline updates.
This change adds snapshot.clipping to all instances of the gridder in 
the pipeline, with user-definable parameters for the continuum & 
spectral-line cases.
It also allows the Selavy job in the self-cal task to take into 
account the weights image and apply a cutoff (to prevent spurious 
artefacts creeping into the selfcal solution).