Pipeline updates for ASKAPSDP-2059 & ASKAPSDP-2060.
This change adds the ability to extract individual beam directions, 
and make images centred at these locations. This is done through 
parsing the output of mslist and footprint.py.
It also introduces the ability to run self-calibration using a 
cmodel-generated model (rather than a components parset). This is 
selectable in a similar way to the continuum-subtraction. 
There are also a few fixes and tidying-up changes to some parts of the 
code.