ASKAPSDP-2079 - Improvements to FITS header metadata.
This change increases the amount of metadata written to the FITS 
header - DATE-OBS, TELESCOP, INTIME & SBID, as well as HISTORY 
statements indicating the time of processing & the versions of 
askapsoft & askappipeline. 
Changes were made to how the metadata is obtained, with a new script 
running before (and calling) the findBeamCentres, to get the metadata 
from the MS via mslist.