ASKAPSDP-2091 - Fixing definition of science MS name.
This change removes the check for the SPLIT or FLAG flags before 
setting the MS_INPUT_SCIENCE name. This way it will get set when not 
given in the config file, even if we are doing no SPLITting or 
FLAGging.
Also fixes the handling of the PROCESS_DEFAULTS_HAS_RUN flag in both 
processDefaults and the defineArtifactsScript file.