ASKAPSDP-2195 - Location of bandpass table.
Fixing parsing of DO_* flags so that we are able to set the location
of the bandpass table within findBandpass.sh. The case of
DO_1934_CAL=false turns off all the cal-related DO flags, but still
goes into each of the cal scripts. This allows us to run the
pre-processing commands including setting the cal table path.

The same thing is done with DO_SCIENCE_FIELD, so that if this is
false, all the science switches are turned off.