ASKAPSDP-2171 - Improving continuum mosaicking
This change overhauls the mosaicking of the continuum images, ensuring 
we get mosaics for each field and an overall mosaic. Doesn't yet 
address the aim of 2171, to get a mosaic per tile, but sets it up.
The code to get the image names has been improved and unified, so that 
there is less repeated code and a unified interface for different 
image types (cont, contcube, spectral).
The getArtifcacts script and the convertFITS jobs also make use of 
these changes, to ensure uniformity.