ASKAPSDP-2157 - Using askapcli/footprint to get beam centres
This change makes use of the footprint tool in the askapcli module to 
obtain the beam centre locations. This is the default path, however we 
still make use of footprint.py (in $ACES) for the case of either BETA 
data or when there is an older observation that has a footprint not 
covered by the CLI system. 
We also only make use of the schedblock tool for ASKAP data (if 
IS_BETA=true - introduced with ASKAPSDP-2296 - we avoid using it).