ASKAPSDP-2173 - Changing the way the continuum sourcefinding is run by the pipeline.
This change makes the default source-finding approach to run on the 
mosaics, including the self-cal mosaics if they are made, and the 
individual beam selavy jobs are only run if the new parameter 
DO_SOURCE_FINDING_BEAMWISE=true. 
The way the image names are set up for the source-finding task has 
been changed a little, to make use of the setImageProperties function.