ASKAPSDP-2337 - Fixing source-finding script
This change ensures that we always use the correct image when running 
the sourcefinding, particularly on the continuum mosaics. It also sets 
the weights image correctly, and makes a unique slurm file and working 
directory for each instance (separating out selfcal loops).
Also fixing a few instances where the ` character was not escaped 
correctly (ie. need \` instead of just `)