ASKAPSDP-2347 - Addressing footprint determination corner case
This change moves the field determination from findBeamCentres to 
prepareMetadata. This ensures that it is always done, even when we 
don't need to find the beam centre positions. 
Also add a condition to the call within sciencecalIm to the 
getBeamOffsets function, so that we don't do it if we haven't worked 
out the beam locations.
Added additional comments in various places as well, to make things a 
bit more clear.