ASKAPSDP-2385 - Improving the use of Stokes I model in RM Synthesis
This change ultimately fixes the FDF-normalisation issue of 
ASKAPSDP-2385, but also improves the way we handle the Stokes I model. 
I've introduced a new class to handle the model, that allows switching 
between fitting a polynomial or using the Taylor-term information to 
parameterise the Stokes I spectrum. This has flowed through to a 
number of improvements to the code in associated classes for RM 
synthesis.