ASKAPSDP-2414 - Separating linmos & sourcefinding jobs a little.
This change makes a clearer separation between the linmos jobs and the 
source-finding jobs on the mosaics. The calls to the source-finding 
tasks are now done from a separate file runMosaicSourcefinding, that 
handles both cases of mosaicking. 
There's a bit of cleaning up of the altLinmos script, as well as 
removing the FIELDBEAMJOB variable from everywhere except for the 
setJob funtion in utils (the only place where it is needed).