ASKAPSDP-2144 - Defining artefacts and the casda upload script.
This change allows the defineArtefacts script to set the wildcard 
names for the spectra & moment map elements that are incorporated into 
the CASDAupload parset. The types are hard-coded in the casda.sh 
script, but the names are set only when the respective selavy 
directory is defined. There are essentially two cases - either 
spectral-line (spectra, noise spectra & moment maps), or polarisation 
(spectra & noise spectra for each stokes, plus FDF & RMSF for the Q 
cube only).