ASKAPSDP-2140 - FITS conversion for extracted spectral artefacts.
This change adds the ability to convert all extracted spectra (& 
spectra-like) artefacts to FITS format. This is done at the end of the 
respective Selavy jobs, calling the FITS-convert script in turn. We 
don't save an individual parset each time, and the log just gets 
appended (this is the usual case for the fitsConvertText now). 

The definition of the fitsConvertText has been moved from a funtion in 
utils.sh to processDefaults.sh, so that it gets defined once at the 
start and there is no subsequent need to call any function. The parset 
and log filenames have to be set before calling fitsConvertText, and 
the affected code has been updated.