ASKAPSDP-2144 - Removing <project> from derived types and improving code structure.
This solves the principal problem arising from ASKAPSDP-2472 of there 
being an erroneous <project> tag in SpectrumElement & 
MomentMapElement. 
At the same time, to simplify the code a bit, I've made another base 
class DerivedElementBase, that incorporates the bulk of the 
functionality of these two classes, and allows them to derive from it.