ASKAPSDP-2144 - Fixing inheritence in the casdaupload classes.
Have changed DerivedElementBase to inherit directly from ElementBase, 
with the addition of the type handling that is in TypeElementBase. 
Since we don't have <project>, we can't derive from 
ProjectElementBase.
Also added code in ImageElement to put in the artifactparam name into 
the parset being used to construct the spectra/momentmaps, as this is 
required for certain error messages.