ASKAPSDP-2503 - Further improvements to thumbnail task
This adds the ability to plot the locations of continuum components on
the thumbnail image, along with the weights contours. This is
envisaged as an additional output that has been added to the pipeline
task - each continuum image (of type cont_restored_T0) that has an
associated component catalogue gets one of these enhanced thumbnails.