ASKAPSDP-2543 - Initial version of diagnostics script for pipelines
This adds a prototype pipeline script that creates some basic 
diagnostic plots - images of the 2D data with catalogues overlaid 
(which were in makeThumbnails but have now been removed), and PNGs of 
the noise maps created by selavy.
Minor update to makeThumbnailImage.py that allows an output directory 
to be specified.
There is a new directory in the pipeline output, called 'diagnostics'.