ASKAPSDP-2508 - Fixes for the case where the SB MS is absent.
This adds more robustness to the prepareMetadata script, so that if 
the SB MS is not present (ie. it has been purged) but a user wants to 
continue the processing that had been started when it was present, the 
user can specify the MS name and it will identify a previously-made 
metadata file. 

If the derived MSs (the split full-resolution one or the averaged one 
in the science-field case) are both absent, the relevant processing 
will be turned off.