ASKAPSDP-2590 - A bunch of changes aimed at getting the noise maps included in the archived data products.
This adds the noise maps to the list of desired artefacts. Doing so 
entailed streamlining the way we name things that Selavy uses - the 
noise maps & related images, and the directories. These are now 
defined in the utils.sh script.

In doing so I also identified and fixed an issue with the dependencies 
based on the continuum cube linmos jobs.