ASKAPSDP-2643 - Adding beam logging to new imager
The askap_imager is now able to write out a beam log when run in 
spectral-line mode (specifically, when solverpercore=true).
The name of the beam log is hard-coded to match the image cube name, 
so that if the image cube is image.i.blah, the beam log will be 
beamlog.image.i.blah.txt. Multiple output cubes mean multiple beam 
logs.