ASKAPSDP-2473 - Further changes to help distributed parameterisation.
First development of the distributed RM synthesis, although not yet 
implemented. New class building on the base distributed parameteriser. 
To get this working, needed to make CasdaPolarisationEntry capable of 
being sent via blobs. 
Changed ValueError to be just a double class (not templated), and had 
to change the name of casda.h to Casda.h (since I made a .cc to go 
with it, and the lower-case version doesn't get picked up by rbuild).
Also added an id() function to CasdaPolarisationEntry, along with 
default constructors for it and ComponentCatalogue