ASKAPSDP-2473 - Initial version of distributed RM synthesis.
Have got a version of the distributed RM synthesis, using the 
DistributedParameteriserBase method, that appears to work, at least as 
far as building & running without error, and distributing the 
processing across nodes. Now need to check for correctness of the 
result.
Note that there are a bunch of extra log messages that will need to be 
removed. Also, the interface for passing cubes around should be 
settled on - currently have a mix of references & pointers.