ASKAPSDP-2473 - Fixing up the HI parameterisation.
There were a couple of parameters that were not being calculated 
properly in the distributed HI processing. 
First, the flags were not being passed via the blob.
Second, the object noise level was never calculated for the 
distributed parameterisation, so the integFluxError was failing. (this 
was a problem even for the trunk version).
We now ensure the initial distributed parameterisation (done by 
DistributedFitter) does the noise calculation & related things even 
when fitting is not being done.