ASKAPSDP-2589 - Allowing output to FITS of noise maps etc
This change allows the ImageWriter (and hence DistributedImageWriter) 
classes to write directly to FITS. This makes use of the FITSImageRW 
accessor class, and requires the addition of parsets to the interface.
The accessor class has been improved to allow the writing of an 
arbitrary subset (rather than requiring the full spatial extent). This 
makes it consistent with the CASA interface.