ASKAPSDP-2701 - HI parameterisation improvements
This adds an HIdata class that handles, for now, the extraction of all 
spectra, maps & cubelets for a given HI source. It allows us to 
control the parameterisation of the extractors, so that they match 
what we want for the HI sources.
Also a minor change to the logging in CubeletExtractor.