ASKAPSDP-2712 - Better module loading/unloading within slurm jobs
For modules other than the askapsoft ones, we've now got a more 
transparent & consistent way of handling them within the slurm jobs. 
There are new functions loadModule/unloadModule that will load a 
named module only when it is not already loaded, and be able to unload 
it for the situation where it was just loaded. 

Also fixing a bug with the way the bandpass smoothing script was being 
called, removing the quotes from around the script arguments.