ASKAPSDP-2592 - Production of a component map & residual
This change alters the way the continuum catalogue processing is done. 
The construction of the island & component catalogues is now done in a 
distributed fashion, together with the creation of a component map 
(ie. the brightness of just the fitted components), as well as a 
residual map (input image - component map). 
It is integrated into Selavy - the images produced are 
componentMap_imagename and componentResidual_imagename.