ASKAPSDP-2589 - Supporting FITS output by allowing update within the pipeline of FITS headers
This adds a new python script to the pipeline directory, that uses 
astropy's pyfits library to update a fixed set of headers. The FITS 
conversion code has been changed so that this is always done, even if 
the FITS file exists. This catches newly-created files, particularly 
those out of Selavy where we don't do the conversion on them.