ASKAPSDP-2801 - Fixing the PSF fitting
This fixes a bug related to rev.8901, where the support size used in 
fitting to the PSF was altered to ensure it was always odd. 
The (fixed) location of the peak of the Gaussian is set to 
newShape[0]/2 - we now use integer division rather than float (ie not 
newShape[0]/2.), so that we get the centre pixel of the support array.