ASKAPSDP-2789 - Allowing calculation of Island statistics.
This adds calculations to obtain the background, the noise, and the 
stats on the fit residual for each island. Use extractors to get the 
image/mean/noise map pixel values, and subtract off the fitted 
Gaussian fluxes to get the residuals. 
Needed one small change to the SourceDataExtractor to handle multiple 
Stokes correctly.
Some formatting for the committed files as well.