ASKAPSDP-2842 - Imcontsub interface in pipeline scripts
This updates the interface to the image-based continuum-subtraction 
scripts as used by the pipeline. This makes use of the new 
command-line interfaces, and allows access to the contsub_im.py script 
as well as robust_contsub.py. There are a bunch of new parameters 
available for each of the possible command-line arguments.